requirements.txt
# Core dependencies
discord.py>=2.0.0
aiohttp>=3.8.0
twitchio>=2.6.0
python-dotenv>=0.19.0
requests>=2.28.0
pymongo[srv]>=4.6.0  # MongoDB with DNS SRV support

# Advanced Text Processing (optimized for Render free tier)
spacy>=3.4.0,<4.0.0
nltk>=3.8.0
textblob>=0.17.0  # Simple NLP tasks and sentiment
pattern

# Deployment and Web
fastapi>=0.95.0
uvicorn[standard]>=0.21.0
gunicorn>=20.1.0
python-json-logger>=2.0.0
aiohttp-cors>=0.7.0  # Added for health check server

# Monitoring & Storage
motor>=3.3.0  # Async MongoDB driver
psutil>=5.9.0

# AI & Embeddings
google-generativeai
google-genai  # For new Gemini API
sentence-transformers
numpy
